remove unneeded int cast

+++ src/main/kotlin/gay/vaskel/aoc2021/days/DayFour.kt
@@ -9,7 +9,7 @@ class DayFour: BaseDay {
         fun getSum(nums: List<Int>): Int {
             return rows.sumOf { lst ->
                 lst.sumOf {
-                    if (!nums.contains(it)) it else 0 as Int
+                    if (!nums.contains(it)) it else 0
                 }
             }
         }
